+++ src/flip-queries/dto/add-flip-query.dto.ts
