+++ src/flip-queries/interface/queries.interface.ts
