+++ src/tools/workingWithFile.ts
@@ -1,9 +1,9 @@
-import * as path from 'path';
-import * as fsPromises from 'fs/promises';
-import { CardTypes } from '../card-poe-data/interface/card-types';
-import { PoeTradeDataItemsResponse } from '../types/response-poe-fetch';
-import { QueriesItems } from '../poe-queries/interface/queries.interface';
 import * as fs from 'fs';
+import * as fsPromises from 'fs/promises';
+import * as path from 'path';
+import { CardTypes } from '../card-poe-data/interface/card-types';
+import { QueriesItems } from '../flip-queries/interface/queries.interface';
+import { PoeTradeDataItemsResponse } from '../types/response-poe-fetch';
 
 export enum fileNamesEnum {
   POE_QUERIES_SEARCH = 'poeSearchUrls.json',
